update DrawCharacterIcon and DecalHUD_DrawPolyGT4 (not a rewrite)

+++ psx-modding-toolchain/games/CrashTeamRacing/decompile/General/DecalHUD/DecalHUD_DrawPolyGT4.c
@@ -1,6 +1,6 @@
 #include <common.h>
 
-void DECOMP_DecalHUD_DrawPolyGT4(struct Icon* icon, u_int posX, int posY, struct PrimMem* primMem, u_int* ot, u_int topLeftColor, u_int topRightColor, u_int bottomLeftColor, u_int bottomRightColor, u_char semitransparencyEnabled, short scale)
+void DECOMP_DecalHUD_DrawPolyGT4(struct Icon* icon, u_int posX, int posY, struct PrimMem* primMem, u_long* ot, u_int topLeftColor, u_int topRightColor, u_int bottomLeftColor, u_int bottomRightColor, u_char semitransparencyEnabled, short scale)
 {
 	u_char y2;
 	u_short bottomRightCorner;
